+++ keep/src/components/CardList/index.js
@@ -1,16 +1,36 @@
-import React,{useState } from 'react';
+import React, { useEffect, useState,useContext } from 'react';
 import PropTypes from 'prop-types';
 import {db} from '../../firebase-config'
-import {updateDoc, doc} from 'firebase/firestore'
+import {collection, getDocs, where, query,doc,updateDoc} from 'firebase/firestore'
+import {AuthContext} from '../../Auth';
 import NoteCard from '../NoteCard/index';
 import NotesInput  from '../NotesInput/index';
 import EditNotes from '../EditNotes';
 import './style.css'
 
-const CardList=({view,notes,setIsUpdate}) =>{
+const CardList=({view}) =>{
     const [editModle,setEditModle]=useState(false);
     const [updateNote,setUpdateNote]=useState({});
+    const [isUpdate,setIsUpdate]=useState(false);
+    const [notes,setNotes]=useState([]);
 
+    const {user}=useContext(AuthContext)
+
+  useEffect(()=>{
+    const fetchData=async()=>{
+      try{
+        const noteColloectionRef=doc(db,'Users',user.uid)
+        const subCollectionRef=collection(noteColloectionRef,'Notes')
+        const notesquery=query(subCollectionRef,where("isDeleted","==",false))
+        const data =await getDocs(notesquery);
+        setNotes(data.docs.map((doc)=>({...doc.data(),id:doc.id })))
+      }catch(e){
+        console.log(e)
+      }
+    }
+    fetchData()
+
+  },[isUpdate])
     const displayCardContent=(item)=>{
         setEditModle(true)
         setUpdateNote(item)
@@ -54,8 +74,6 @@ const CardList=({view,notes,setIsUpdate}) =>{
 }
 CardList.propTypes = {
  view:PropTypes.string.isRequired,
- notes:PropTypes.arrayOf.isRequired,
- setIsUpdate:PropTypes.func.isRequired
   };
 
 export default CardList;
